Use correct key when hashing photos
I introduced a bug recently that assumed we were using `four11_key` instead of `session_key` to hash student photos, which meant that the new 2022 photos did not display even after being uploaded to the storage bucket as their filenames were wrong.

This fixes the photo script to use the correct key.