Simple runtime edit
The issue is that in past commits (#648 and #649) there were errors with being unable to deploy to app engine. The errors appears to have been with deprecated support for python 3.7, the version we currently use.

I changed it to 3.12 to hopefully not have to deal with errors for a while. Testing it locally appears to work, although I'm not sure if it is reflected when testing locally. Also not sure how to actually test this error without going for the commit

More information on deprecated runtimes are here: https://cloud.google.com/appengine/docs/standard/lifecycle/support-schedule